--- a/compile/compile_bench_test.go
+++ b/compile/compile_bench_test.go
@@ -19,30 +19,19 @@ func BenchmarkCompileDynamicPolicy(b *testing.B) {
 	//ctx := context.Background()
 
 	numPolicies := []int{1000, 2500, 5000, 7500, 10000}
+	testcases := map[int]map[string]string{}
+
+	for _, n := range numPolicies {
+		testcases[n] = generateDynamicPolicyBenchmarkData(n)
+	}
+
+	b.ResetTimer()
 
 	for _, n := range numPolicies {
 		b.Run(fmt.Sprintf("%d", n), func(b *testing.B) {
-			files := map[string]string{
-				"main.rego": `
-					package main
-
-					denies[x] {
-						x := data.policies[input.type][input.subtype][_].denies[_]
-					}
-					any_denies {
-						denies[_]
-					}
-					allow {
-						not any_denies
-					}`,
-			}
-
-			for i := 0; i < n; i++ {
-				files[fmt.Sprintf("policy%d.rego", i)] = generateMockPolicy(i)
-			}
-
-			test.WithTempFS(files, func(root string) {
+			test.WithTempFS(testcases[n], func(root string) {
 				b.ResetTimer()
+
 				compiler := New().
 					WithPaths(root)
 
@@ -87,6 +76,29 @@ func BenchmarkCompileDynamicPolicy(b *testing.B) {
 			// }
 		})
 	}
+}
+
+func generateDynamicPolicyBenchmarkData(N int) map[string]string {
+	files := map[string]string{
+		"main.rego": `
+			package main
+
+			denies[x] {
+				x := data.policies[input.type][input.subtype][_].denies[_]
+			}
+			any_denies {
+				denies[_]
+			}
+			allow {
+				not any_denies
+			}`,
+	}
+
+	for i := 0; i < N; i++ {
+		files[fmt.Sprintf("policy%d.rego", i)] = generateMockPolicy(i)
+	}
+
+	return files
 }
 
 func generateMockPolicy(N int) string {