This PR is intended to address some of the performance issues brought to light in issue #5216.

The Compile API's performance nosedives surprisingly quickly when working with large policy sets (>10k policy files), especially when policy selection at the top-level depends on a `Ref` type (such as something from the `input` document).

## What the PR includes:

- New benchmark for the `compile` package (just builds a large policy set, attempts to run `Build()` on them)
- Optimizations to the type checker (a linear scan switched to binary search, 3-5x speedup for benchmark)
- Miscellaneous for loop optimizations in `ast/visit`, resulting in a 10-15% memory usage improvement.

## Task List:

- [X] Investigate memory usage blowup at ~2500 policies in benchmark (looks like a case where the Golang slice allocator is breaking down, similar to what happened for sets/objects in older OPA versions).
  - **Result**: This is caused by excessive slice copying in the [`Merge` function](https://github.com/open-policy-agent/opa/blob/f7ee81577cac1d0d5634a42db0b49e48c8d6fd40/types/types.go#L471-L486) in `ast/types`. It's a similar problem to the sorting of Set/Object key slices that came up a month or two back, and could probably be resolved similarly by making the sorting lazier, or using a more efficient underlying set-like data representation. The main problem is that the assumption that `Any` is a slice of types is baked into several places in the compiler, and is part of the public interface of the `types` package. I don't see a way to resolve the massive amount of copying except through a big (breaking) data structure change.

## Benchmark Charts

Below is chart comparing the work done in this PR against what's on `main` right now:

![2022-10-26_5216-benchmark-results](https://user-images.githubusercontent.com/1906841/198153201-c3291a89-7e82-49e1-89f2-d946ce29421b.png)

As you can see in the above image, the work of this PR shifts the "performance nosedive zone" to the right a bit (explodes above 2500 policies, instead of at <2500 policies). However, this PR *significantly improves the runtime of the benchmark across all input sizes, transforming an exponential curve into a flatter, logarithmic curve.

For more the actual benchmark stats / profiling charts behind the graphs, see [this comment](https://github.com/open-policy-agent/opa/pull/5307#issuecomment-1292739806) later in the discussion thread.