@philipaconrad 2022-10-26
I discovered today while I was looking over `ast/visit.go` that many of the `Walk` family of functions there use the idiom `for _, element := range someCollection { ... }`, which [generates extra element copies](https://github.com/teivah/100-go-mistakes#ignoring-that-elements-are-copied-in-range-loops-30).

Switching all of those loops to use index-based collection lookups results in about a 10-15% memory traffic reduction in the compile benchmark, and a dramatic performance improvement in actually running the benchmark (50+ seconds per run -> ~30 second per run). I believe the speedups to be mostly to reduced copying/GC pressure.