@philipaconrad 2022-10-26
This PR doesn't fully address the issues brought up in #5216, so I'm not entirely comfortable marking it as "fixes <issue number>" for that issue.

It may be worth re-branding this PR as type checker speedups, since that's 90% of what was improved here.