@philipaconrad 2022-10-26
The benchmarks for the new `BenchmarkCompileDynamicPolicy` benchmark test:

**Note**: _When reading the graphs, pay attention to the overall stats in the grey block at the top! The memory profiling graphs have different overall amounts of traffic due to performance differences (7+ GB for `main`, 15+ GB for this PR), and this directly affects the amount of memory traffic shown in the graphs going to each node._

### `main` (v0.45.0, commit da8afed):
```
go test -benchmem -run=^$ -tags opa_wasm,slow -bench ^BenchmarkCompileDynamicPolicy$ github.com/open-policy-agent/opa/compile
goos: linux
goarch: amd64
pkg: github.com/open-policy-agent/opa/compile
cpu: Intel(R) Core(TM) i7-8650U CPU @ 1.90GHz
BenchmarkCompileDynamicPolicy/1000-8         	1000000000	         0.2856 ns/op	       0 B/op	       0 allocs/op
BenchmarkCompileDynamicPolicy/2500-8         	       1	1162606229 ns/op	315331832 B/op	 3546221 allocs/op
BenchmarkCompileDynamicPolicy/5000-8         	       1	3714670878 ns/op	950775240 B/op	 7087389 allocs/op
BenchmarkCompileDynamicPolicy/7500-8         	       1	8569930462 ns/op	1889546472 B/op	10631323 allocs/op
BenchmarkCompileDynamicPolicy/10000-8        	       1	18399650764 ns/op	3121444856 B/op	14175778 allocs/op
PASS
ok  	github.com/open-policy-agent/opa/compile	37.956s
```
<details>
  <summary><b><u>CPU Profiling Graph</u></b></summary>
  <img src="https://user-images.githubusercontent.com/1906841/198150035-cbae8107-a14f-4c6f-8e47-b0b97cdef2f0.png" />
</details>
 <details>
  <summary><b><u>Memory Profiling Graph</u></b></summary>
<img src="https://user-images.githubusercontent.com/1906841/198150295-e4e3bfeb-8f82-4efe-84df-7283a3fcdd61.png" />
</details>
 

### This PR:
```
go test -benchmem -run=^$ -tags opa_wasm,slow -bench ^BenchmarkCompileDynamicPolicy$ github.com/open-policy-agent/opa/compile
goos: linux
goarch: amd64
pkg: github.com/open-policy-agent/opa/compile
cpu: Intel(R) Core(TM) i7-8650U CPU @ 1.90GHz
BenchmarkCompileDynamicPolicy/1000-8         	1000000000	         0.2126 ns/op	       0 B/op	       0 allocs/op
BenchmarkCompileDynamicPolicy/2500-8         	1000000000	         0.5700 ns/op	       0 B/op	       0 allocs/op
BenchmarkCompileDynamicPolicy/5000-8         	       1	1362136949 ns/op	950682232 B/op	 7086254 allocs/op
BenchmarkCompileDynamicPolicy/7500-8         	       1	2385335304 ns/op	1889528904 B/op	10635632 allocs/op
BenchmarkCompileDynamicPolicy/10000-8        	       1	3507656799 ns/op	3121244984 B/op	14174250 allocs/op
PASS
ok  	github.com/open-policy-agent/opa/compile	30.960s
```

Profiling Graphs:

<details>
  <summary><b><u>CPU Profiling Graph</u></b></summary>
  <img src="https://user-images.githubusercontent.com/1906841/198150634-06639fce-5372-4eec-bd25-e289b8aadd1a.png" />
</details>
<details>
  <summary><b><u>Memory Profiling Graph</u></b></summary>
<img src="https://user-images.githubusercontent.com/1906841/198150664-0fe3818e-784f-4992-940e-e0265f9800eb.png" />
</details>