@philipaconrad 2022-10-26
@srenatus I'm marking this PR as **RFR**, so that we can get the performance improvements for the type checker in, and then worry about dramatic refactorings of the type system in a future PR. I've re-scoped the PR title accordingly.